Completed 6.72

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,12 +1,33 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
+    async loginUser(parent, args, { prisma }, info) {
+        // console.log('args: ', args)
+        const user = await prisma.query.user({
+            where: {
+                email: args.email
+            }
+        })
+        if (!user) {
+            throw new Error('Unable to login with these credentials')
+        } 
+        
+        const pwIsMatched = await bcrypt.compare(args.password, user.password)
+        if (!pwIsMatched) {
+            throw new Error('Unable to login with these credentials')
+        } else {
+            return {
+                token: jwt.sign({ userId: user.id }, '_my_Sexy_secret_'),
+                user: user
+            }
+        }
+    },
     async createUser(parent, args, { prisma }, info) {
         if (args.data.password.length < 8) {
             throw new Error('Password must be at least 8 characters long')
         }
-
         const password = await bcrypt.hash(args.data.password, 10)
         const user = await prisma.mutation.createUser({ 
             data: {
@@ -14,36 +35,49 @@ const Mutation = {
                 password
             }
         })
-
         return {
             user,
             token: jwt.sign({ userId: user.id }, '_my_Sexy_secret_')
         }
     },
-    deleteUser(parent, args, { prisma }, info) { 
+    deleteUser(parent, args, { prisma, request }, info) { 
+        const userId = getUserId(request)
         return prisma.mutation.deleteUser({ 
-            where: { id: args.id }
+            where: { id: userId }
         }, info)
     },
-    updateUser(parent, args, { prisma }, info) {
+    updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.updateUser({
-            where: { id: args.id },
+            where: { id: userId },
             data: args.data
         }, info)
     },
-    createPost(parent, args, { prisma }, info) {
+    createPost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createPost({
             data: {
                 title: args.data.title,
                 body: args.data.body,
                 published: args.data.published,
                 author: {
-                    connect: { id: args.data.author}
+                    connect: { id: userId }
                 }
             }
         }, info)
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if (!postExists) {
+            throw new Error('Unable to delete post')
+        }
         return prisma.mutation.deletePost({
             where: { id: args.id }
         }, info)
